Remove ghost file from project filters
The "carbon_cavern.ogg" file was a file containing a fanmade soundtrack.
I never got permission to use it, so I decided to remove it when making
the source code for this project public (I never got to a point where I
could use it anyway, sadly). However, when doing so, I forgot to remove
the file from the filters.

Also, the "Carbon Cavern" track is absolutely fantastic. You should
definitely give it a listen!: https://youtu.be/SSlYzCb4618